+++ sdk/spring/azure-spring-cloud-stream-binder-eventhubs/CHANGELOG.md
@@ -3,6 +3,9 @@
 ## 2.2.0-beta.1 (Unreleased)
 
 
+## 2.1.0 (2021-01-20)
+
+
 ## 2.0.0 (2020-12-30)
 ### Breaking Changes
 - Deprecated the `spring.cloud.azure.managed-identity.client-id` property,